Fix issue with two midrolls in one cue point
